refactor: move meta generation from layout to inside page enable dynamic.
