+++ src/app/friends/page.tsx
@@ -3,6 +3,21 @@ import { getConfig } from '@/services/config/getConfig';
 import { getPostData } from '@/services/content/posts';
 import '@/styles/friendsLinks.css';
 import { PostData } from '@/types';
+import type { Metadata } from 'next';
+
+export async function generateMetadata(): Promise<Metadata> {
+  const config = getConfig();
+  return {
+    title: `Friends - ${config.title}`,
+    description: `Friends page of ${config.title} - ${config.description}`,
+    openGraph: {
+      title: `Friends - ${config.title}`,
+      description: `Friends page of ${config.title} - ${config.description}`,
+      type: 'website',
+      locale: config.lang,
+    },
+  };
+}
 
 export default async function FriendsPage() {
   const post: PostData = await getPostData('Friends', 'Friends');
